python-service/requirements.txt
fastapi
uvicorn
langgraph
zep-python
langchain-openai
python-dotenv
pydantic-ai
openai
httpx 